backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Rishikesh Maddhesiya <rishi89324@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = ">=0.26.0"
langgraph = "0.2.59"
pydantic = "2.7.4"
python-dotenv = "0.21.0"
langchain = "0.3.12"
langchain-community = "0.3.12"
fastapi = "0.115.2"
langchain-anthropic = "0.2.3"
openai = "1.55.3"
copilotkit = "0.1.30"
langgraph-cli = "^0.1.65"
html2text = "^2024.2.26"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
